rename: fetch 파일명을 백엔드 도메인 네이밍과 동일하도록 수정
