+++ src/recoil/couponList/couponListState.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 import type { Coupon } from '../../types/coupon.type';
-import { requestCouponList } from '../../apis/requests/couponList';
+import { requestCouponList } from '../../apis/requests/coupon';
 
 export const couponListState = atom<Coupon[]>({
   key: 'couponListState',
